Refactor dev-middleware as embedded plugin
