+++ lib/jamsite-pages-dev.js
@@ -1,3 +1,4 @@
+const path = require('path')
 const { JamsitePages } = require('@jamsite/jamsite')
 
 const EVENT_UPDATE = 'update'
@@ -7,6 +8,12 @@ class JamsitePagesDev extends JamsitePages {
     const config = super.defaultConfig()
     config.chokidarOptions.persistent = true
     return config
+  }
+
+  initPlugins () {
+    const plugins = super.initPlugins()
+    plugins.loadPluginsFromDir(path.join(__dirname, 'plugins'))
+    return plugins
   }
 
   bindWatcher (watcher) {
